backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.25.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.10.0

# Authentication
authlib==1.3.0
itsdangerous==2.1.2
httpx>=0.28.1

# AutoGen dependencies
autogen-agentchat>=0.4.2
openai>=1.0.0
google-genai==1.41.0